Relocate constants folder
