+++ src/constants/dbErrors.ts
